TestStudio/requirements.txt
DateTime==4.3
Django==2.2
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
Markdown==3.3.6
pick==0.6.4
pytz==2021.3
questionnaire==2.2.0
sqlparse==0.4.2
typing-extensions==4.0.1
zipp==3.6.0
zope.interface==5.4.0
